Create project-plan-draft-11-22-2024
This is an updated project plan based on the Google Doc. This is a much better format. I also added some items based on suggestions I read about (without knowing much about them). Feel free to delete or add anything. Note the dropdown description boxes. This seemed like an easy way to add initial notes for these items that could be useful in the project doc.